convert naming provider to use a pstore
The existing method of using flat text files meant that indices had to
be manually updated. This was causing terribleness in the form of nil
references and infinite loops.

A rake task was added to use the text files to generate a binary
pstore indexed by letter name. This is more efficient since it is
thunked more quickly into memory and names are reached in constant time

fixes #28